[conf] Update the ovs_events config for Apex
The SendNotification option in ovs_events is set to
false in Apex templates. This means that no notifications
will be emitted on link status change.

This patch updates ovs_events.conf.erb to match the
config in the samples folder, adn the ones used by
the reference container

Change-Id: Ia034b81aedc67c2ced51cbf831baca176218d7d3
JIRA: BAROMETER-91
Signed-off-by: Emma Foley <emma.l.foley@intel.com>